requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch~=1.11.0
torchvision~=0.12.0
Pillow~=8.4.0
fastapi==0.76.0
pydantic==1.9.0
python-dotenv==0.20.0
ImageHash==4.2.1
pytest==7.1.2
uvicorn==0.17.6
numpy~=1.22.3
opencv-python~=4.5.5.64
imutils~=0.5.4
scipy~=1.7.3